fix: resolve all lint errors in objectbrick schema mapping hooks
- Extract static helper functions (filterTreeNodesRecursive,
  collectAllKeys, scanForObjectBricksFields, removeObjectBricksNodes)
  into class-attributes-tree-helpers.ts to bring use-class-attributes-tree
  under the 300-line max-lines limit
- Cast fieldDefinition.fieldtype to string to fix unsafe-argument error
- Add missing useMemo deps (allBrickKeys, fieldDefinitionRegistry, t)
  and remove invalid eslint-disable comment for non-existent
  react-hooks/exhaustive-deps rule
- Cast responses[index].data to ConfigLayoutDefinition to fix
  unsafe-argument error in use-objectbrick-layouts
- Add dispatch to useEffect deps and remove invalid eslint-disable
  comment in use-objectbrick-layouts
